Separate test for single_number
